requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
astor==0.8.1
    # via -r requirements.in
black==24.10.0
    # via -r requirements.in
build==1.2.2.post1
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2024.8.30
    # via
    #   -r requirements.in
    #   requests
cfgv==3.4.0
    # via
    #   -r requirements.in
    #   pre-commit
charset-normalizer==3.4.0
    # via
    #   -r requirements.in
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   -r requirements.in
    #   build
    #   click
contourpy==1.3.0
    # via
    #   -r requirements.in
    #   matplotlib
customtkinter==5.2.2
    # via -r requirements.in
cycler==0.12.1
    # via
    #   -r requirements.in
    #   matplotlib
darkdetect==0.8.0
    # via
    #   -r requirements.in
    #   customtkinter
distlib==0.3.9
    # via
    #   -r requirements.in
    #   virtualenv
filelock==3.16.1
    # via
    #   -r requirements.in
    #   virtualenv
flake8==7.1.1
    # via -r requirements.in
fonttools==4.54.1
    # via
    #   -r requirements.in
    #   matplotlib
identify==2.6.1
    # via
    #   -r requirements.in
    #   pre-commit
idna==3.10
    # via
    #   -r requirements.in
    #   requests
imageio==2.36.0
    # via
    #   -r requirements.in
    #   scikit-image
isort==5.13.2
    # via -r requirements.in
kiwisolver==1.4.7
    # via
    #   -r requirements.in
    #   matplotlib
lazy-loader==0.4
    # via
    #   -r requirements.in
    #   scikit-image
matplotlib==3.9.2
    # via -r requirements.in
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   flake8
mouseinfo==0.1.3
    # via
    #   -r requirements.in
    #   pyautogui
mss==9.0.2
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
networkx==3.4.1
    # via
    #   -r requirements.in
    #   scikit-image
nodeenv==1.9.1
    # via
    #   -r requirements.in
    #   pre-commit
numpy==2.1.2
    # via
    #   -r requirements.in
    #   contourpy
    #   imageio
    #   matplotlib
    #   opencv-python
    #   pyclick
    #   scikit-image
    #   scipy
    #   tifffile
opencv-python==4.10.0.84
    # via -r requirements.in
packaging==24.1
    # via
    #   -r requirements.in
    #   black
    #   build
    #   customtkinter
    #   lazy-loader
    #   matplotlib
    #   scikit-image
pathspec==0.12.1
    # via
    #   -r requirements.in
    #   black
pillow==11.0.0
    # via
    #   -r requirements.in
    #   imageio
    #   matplotlib
    #   pyscreeze
    #   scikit-image
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via
    #   -r requirements.in
    #   black
    #   virtualenv
pre-commit==4.0.1
    # via -r requirements.in
pyautogui==0.9.54
    # via
    #   -r requirements.in
    #   pyclick
pyclick==0.0.2
    # via -r requirements.in
pycodestyle==2.12.1
    # via
    #   -r requirements.in
    #   flake8
pyflakes==3.2.0
    # via
    #   -r requirements.in
    #   flake8
pygetwindow==0.0.9
    # via
    #   -r requirements.in
    #   pyautogui
pymonctl==0.92
    # via
    #   -r requirements.in
    #   pywinctl
pymsgbox==1.0.9
    # via
    #   -r requirements.in
    #   pyautogui
pynput==1.7.7
    # via -r requirements.in
pyparsing==3.2.0
    # via
    #   -r requirements.in
    #   matplotlib
pyperclip==1.9.0
    # via
    #   -r requirements.in
    #   mouseinfo
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
pyrect==0.2.0
    # via
    #   -r requirements.in
    #   pygetwindow
pyscreeze==1.0.1
    # via
    #   -r requirements.in
    #   pyautogui
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   matplotlib
python3-xlib==0.15
    # via
    #   -r requirements.in
    #   pyclick
pytweening==1.2.0
    # via
    #   -r requirements.in
    #   pyautogui
    #   pyclick
pywin32==308
    # via
    #   -r requirements.in
    #   pymonctl
    #   pywinbox
    #   pywinctl
pywinbox==0.7
    # via
    #   -r requirements.in
    #   pywinctl
pywinctl==0.4.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   pre-commit
requests==2.32.3
    # via -r requirements.in
scikit-image==0.24.0
    # via -r requirements.in
scipy==1.14.1
    # via
    #   -r requirements.in
    #   scikit-image
simplejson==3.19.3
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   pynput
    #   python-dateutil
    #   xlib
tenacity==9.0.0
    # via -r requirements.in
tifffile==2024.9.20
    # via
    #   -r requirements.in
    #   scikit-image
tomli==2.0.2
    # via
    #   -r requirements.in
    #   black
    #   build
    #   pip-tools
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   black
    #   pymonctl
    #   pywinbox
    #   pywinctl
urllib3==2.2.3
    # via
    #   -r requirements.in
    #   requests
virtualenv==20.26.6
    # via
    #   -r requirements.in
    #   pre-commit
wheel==0.44.0
    # via pip-tools
xlib==0.21
    # via
    #   -r requirements.in
    #   pyclick

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
